# openapi.yaml
openapi: 3.0.0
info:
  title: 꼬리콥터 API 명세
  description: 꼬리콥터 백엔드 API 명세서입니다. 초기 온보딩, 산책 기능, 코스 및 기록 관리를 위한 API들을 포함합니다.
  version: 1.0.0
servers:
  - url: http://localhost:5000/api/v1
    description: 개발 서버
...
paths:
...
  /users/me/terms:
    post:
      tags:
        - 인증 및 온보딩
      summary: 사용자 약관 동의 상태 저장
      requestBody:
        description: 약관 동의 상태
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                isAgreed:
                  type: boolean
      responses:
        '200':
          description: 약관 동의 상태 저장 성공
...
  /users/me/profile:
    put:
      tags:
        - 인증 및 온보딩
      summary: 사용자 및 반려동물 프로필 정보 등록/업데이트
      requestBody:
        description: 업데이트할 사용자 및 반려동물 프로필 정보
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                dogName:
                  type: string
                dogBreedId:
                  type: integer
                dogBirthYear:
                  type: integer
                dogSize:
                  type: string
                dogImageUrl:
                  type: string
                preferredLocationId:
                  type: integer
      responses:
        '200':
          description: 프로필 업데이트 성공
    get:
      tags:
        - 인증 및 온보딩
      summary: 사용자 및 반려동물 프로필 정보 조회
      responses:
        '200':
          description: 프로필 정보 반환
...
  /users/me/walk-records:
    get:
      tags:
        - 코스 및 기록 관리
      summary: 사용자의 모든 산책 기록 목록 조회
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: size
          in: query
          schema:
            type: integer
            default: 10
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [latest, longest, shortest]
            default: latest
      responses:
        '200':
          description: 산책 기록 목록 반환
...